requirements.txt


fastapi

pydantic
uvicorn 
ultralytics
pillow
opencv-python